feat(android): analytics for achievements earned
