+++ android/app/build.gradle.kts
@@ -106,6 +106,7 @@ dependencies {
 	implementation(libs.androidx.compose.ui.tooling.preview)
 	debugImplementation(libs.androidx.compose.ui.tooling.preview)
 
+	implementation(libs.kotlinx.datetime)
 	implementation(libs.sentry)
 	implementation(libs.sentry.compose)
 }
